requirements.txt
groq
psycopg2-binary
python-dotenv
pandas
numpy 
